model/requirements.txt
# ──────────────────────────────────────────────────────────────────────────────
# Plant Disease Detection API – Python Dependencies
# Install with: pip install -r requirements.txt
# ──────────────────────────────────────────────────────────────────────────────

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database (async)
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Configuration
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic[email]==2.10.3
python-dotenv==1.0.1

# Utilities
python-multipart==0.0.19   # form data / file uploads
email-validator==2.2.0     # Pydantic EmailStr validation

# Development / testing (optional – pin versions for reproducibility)
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1              # async test client for FastAPI
